refactor: separate logic and frontend
