+++ ui/package.json
@@ -10,34 +10,40 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@privy-io/react-auth": "^3.9.0",
-    "@privy-io/wagmi": "^2.1.1",
+    "@privy-io/react-auth": "^3.9.1",
+    "@privy-io/wagmi": "^2.1.2",
     "@tailwindcss/vite": "^4.1.18",
     "@tanstack/react-query": "^5.90.12",
-    "react": "^19.2.0",
-    "react-dom": "^19.2.0",
+    "react": "^19.2.3",
+    "react-dom": "^19.2.3",
     "tailwindcss": "^4.1.18",
-    "viem": "~2.41.2",
+    "viem": "~2.43.3",
     "wagmi": "^3.1.0"
   },
   "devDependencies": {
-    "@eslint/js": "^9.39.1",
-    "@types/node": "^24.10.1",
-    "@types/react": "^19.2.5",
+    "@eslint/js": "^9.39.2",
+    "@types/node": "^25.0.3",
+    "@types/react": "^19.2.7",
     "@types/react-dom": "^19.2.3",
-    "@vitejs/plugin-react": "^5.1.1",
+    "@vitejs/plugin-react": "^5.1.2",
     "daisyui": "^5.5.14",
-    "eslint": "^9.39.1",
+    "eslint": "^9.39.2",
     "eslint-plugin-react-hooks": "^7.0.1",
-    "eslint-plugin-react-refresh": "^0.4.24",
+    "eslint-plugin-react-refresh": "^0.4.26",
     "globals": "^16.5.0",
     "typescript": "~5.9.3",
-    "typescript-eslint": "^8.46.4",
-    "vite": "npm:rolldown-vite@7.2.5"
+    "typescript-eslint": "^8.50.1",
+    "vite": "npm:rolldown-vite@7.3.0"
   },
   "pnpm": {
     "overrides": {
-      "vite": "npm:rolldown-vite@7.2.5"
-    }
+      "vite": "npm:rolldown-vite@7.3.0"
+    },
+    "onlyBuiltDependencies": [
+      "@reown/appkit",
+      "bufferutil",
+      "keccak",
+      "utf-8-validate"
+    ]
   }
 }
